+++ src/year_2024/day_02/part_2.rs
@@ -15,7 +15,7 @@ static SOLVER: SolverMetadata<'static> = SolverMetadata {
     input: super::INPUT,
 };
 
-fn solve(input: &str) -> MyResult<u32> {
+fn solve(input: &str) -> MyResult<u64> {
     parse_and_count(input, with_problem_dampener)
 }
